refactor: fold ThanksReactor into ThankingRobot
Both listeners respond to the same messages; no reason for a separate cog.
ThankingRobot now owns react_to_thanks and the shared _normalize/_is_thanks_to_duckbot helpers.
