+++ tests/cogs/robot/setup_test.py
@@ -1,4 +1,4 @@
-from duckbot.cogs.robot import ThankingRobot, ThanksReactor
+from duckbot.cogs.robot import ThankingRobot
 from duckbot.cogs.robot import setup as extension_setup
 from tests.discord_test_ext import assert_cog_added_of_type
 
@@ -6,4 +6,3 @@ from tests.discord_test_ext import assert_cog_added_of_type
 async def test_setup(bot_spy):
     await extension_setup(bot_spy)
     assert_cog_added_of_type(bot_spy, ThankingRobot)
-    assert_cog_added_of_type(bot_spy, ThanksReactor)
